+++ weldx/core/__init__.py
@@ -1,0 +1,1 @@
+from .math_expression import MathematicalExpression
